+++ src/test/java/io/reactiverse/es4x/test/CommonJSCyclicTest.java
@@ -1,4 +1,4 @@
-package io.reactiverse.es4x.nashorn;
+package io.reactiverse.es4x.test;
 
 import io.reactiverse.es4x.Loader;
 import io.vertx.core.Vertx;
@@ -10,11 +10,13 @@ import org.junit.runners.Parameterized;
 import java.util.Arrays;
 import java.util.List;
 
-import static org.junit.Assert.*;
-import static org.junit.Assume.*;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assume.assumeTrue;
+
+import static io.reactiverse.es4x.test.JS.*;
 
 @RunWith(Parameterized.class)
-public class CommonJSGlobalPollutionTest {
+public class CommonJSCyclicTest {
 
   @Parameterized.Parameters
   public static List<String> engines() {
@@ -24,7 +26,7 @@ public class CommonJSGlobalPollutionTest {
   final String engineName;
   final Loader loader;
 
-  public CommonJSGlobalPollutionTest(String engine) {
+  public CommonJSCyclicTest(String engine) {
     System.setProperty("es4x.engine", engine);
     engineName = engine;
     loader = Loader.create(Vertx.vertx());
@@ -36,13 +38,12 @@ public class CommonJSGlobalPollutionTest {
   }
 
   @Test
-  public void shouldHaveSideEffects() {
-    try {
-      // this test verifies that the pollution of the global context behaves like on node
-      loader.require("./pollution/a.js");
-      fail("should throw");
-    } catch (Exception e) {
-      assertEquals("Error: engine is tainted: b", e.getMessage());
-    }
+  public void shouldHaveTheSameSenseOfAnObjectInAllPlaces() {
+    Object stream = loader.require("./lib/cyclic2/stream.js");
+    assertTrue(isFunction(stream));
+    Object readable = getMember(stream, "Readable");
+    assertTrue(isFunction(readable));
+    Object stream2 = getMember(readable, "Stream");
+    assertTrue(isFunction(stream2));
   }
 }
