+++ src/test/java/io/reactiverse/es4x/test/CommonJsUserModulesTest.java
@@ -1,4 +1,4 @@
-package io.reactiverse.es4x.nashorn;
+package io.reactiverse.es4x.test;
 
 import io.reactiverse.es4x.Loader;
 import io.vertx.core.Vertx;
@@ -12,7 +12,7 @@ import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assume.assumeTrue;
-import static io.reactiverse.es4x.nashorn.JS.*;
+import static io.reactiverse.es4x.test.JS.*;
 
 @RunWith(Parameterized.class)
 public class CommonJsUserModulesTest {
