+++ src/test/java/io/reactiverse/es4x/test/FactoryTest.java
@@ -1,9 +1,10 @@
-package io.reactiverse.es4x.graal;
+package io.reactiverse.es4x.test;
 
 import io.vertx.core.Vertx;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
+import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -12,10 +13,15 @@ public class FactoryTest {
 
   private static final Vertx vertx = Vertx.vertx();
 
+  @BeforeClass
+  public static void beforeClass() {
+    System.setProperty("es4x.engine", "Nashorn");
+  }
+
   @Test(timeout = 10000)
   public void shouldDeployVerticle(TestContext ctx) {
     final Async async = ctx.async();
-    vertx.deployVerticle("graal.js:./verticle.js", deploy -> {
+    vertx.deployVerticle("js:./verticle.js", deploy -> {
       ctx.assertTrue(deploy.succeeded());
       async.complete();
     });
@@ -25,7 +31,7 @@ public class FactoryTest {
   @Test(timeout = 10000)
   public void shouldDeployVerticleWithOnStop(TestContext ctx) {
     final Async async = ctx.async();
-    vertx.deployVerticle("graal.js:./verticle2.js", deploy -> {
+    vertx.deployVerticle("js:./verticle2.js", deploy -> {
       ctx.assertTrue(deploy.succeeded());
       vertx.setTimer(1000L, t -> {
         vertx.undeploy(deploy.result(), undeploy -> {
@@ -40,7 +46,7 @@ public class FactoryTest {
   @Test(timeout = 10000)
   public void shouldDeployVerticleWithoutOnStop(TestContext ctx) {
     final Async async = ctx.async();
-    vertx.deployVerticle("graal.js:./verticle.js", deploy -> {
+    vertx.deployVerticle("js:./verticle.js", deploy -> {
       ctx.assertTrue(deploy.succeeded());
       vertx.undeploy(deploy.result(), undeploy -> {
         ctx.assertTrue(undeploy.succeeded());
