+++ src/test/java/io/reactiverse/es4x/test/LoggerTest.java
@@ -1,8 +1,9 @@
-package io.reactiverse.es4x.nashorn;
+package io.reactiverse.es4x.test;
 
 import io.reactiverse.es4x.Loader;
 import io.vertx.core.Vertx;
-import org.graalvm.polyglot.Value;
+import io.vertx.core.logging.Logger;
+import io.vertx.core.logging.LoggerFactory;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -10,13 +11,13 @@ import org.junit.runners.Parameterized;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 
-import static org.junit.Assert.fail;
 import static org.junit.Assume.assumeTrue;
 
 @RunWith(Parameterized.class)
-public class ProcessTest {
+public class LoggerTest {
+
+  final Logger log = LoggerFactory.getLogger(LoggerTest.class);
 
   @Parameterized.Parameters
   public static List<String> engines() {
@@ -26,7 +27,7 @@ public class ProcessTest {
   final String engineName;
   final Loader loader;
 
-  public ProcessTest(String engine) {
+  public LoggerTest(String engine) {
     System.setProperty("es4x.engine", engine);
     engineName = engine;
     loader = Loader.create(Vertx.vertx());
@@ -37,21 +38,22 @@ public class ProcessTest {
     assumeTrue(loader.name().equalsIgnoreCase(engineName));
   }
 
-  @Test(timeout = 10000)
-  public void testProcessEnv() throws Exception {
-    Object res = loader.eval("process.env");
-    Map<String, String> env;
-    if (res instanceof Map) {
-      env = (Map) res;
-    } else if (res instanceof Value) {
-      env = ((Value) res).as(Map.class);
-    } else {
-      fail("Cannot cast response object!");
-      return;
+  @Test
+  public void shouldThrowAnErrorIfFileCantBeFound() {
+    try {
+      throw new RuntimeException();
+    } catch (RuntimeException e) {
+      log.error("Oops!", e);
     }
+  }
 
-    for (Map.Entry<String, String> kv : env.entrySet()) {
-      System.out.println(kv.getKey() + ": " + kv.getValue());
+  @Test
+  public void shouldPrettyPrintException() {
+    Object module = loader.require("./exp.js");
+    try {
+      loader.invokeMethod(module, "a");
+    } catch (RuntimeException e) {
+      log.error("Error from Script", e);
     }
   }
 }
