+++ src/test/java/io/reactiverse/es4x/test/ProcessTest.java
@@ -1,24 +1,22 @@
-package io.reactiverse.es4x.nashorn;
+package io.reactiverse.es4x.test;
 
 import io.reactiverse.es4x.Loader;
 import io.vertx.core.Vertx;
-import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.LoggerFactory;
+import org.graalvm.polyglot.Value;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.Map;
 
+import static org.junit.Assert.fail;
 import static org.junit.Assume.assumeTrue;
 
 @RunWith(Parameterized.class)
-public class LoggerTest {
-
-  final Logger log = LoggerFactory.getLogger(LoggerTest.class);
+public class ProcessTest {
 
   @Parameterized.Parameters
   public static List<String> engines() {
@@ -28,7 +26,7 @@ public class LoggerTest {
   final String engineName;
   final Loader loader;
 
-  public LoggerTest(String engine) {
+  public ProcessTest(String engine) {
     System.setProperty("es4x.engine", engine);
     engineName = engine;
     loader = Loader.create(Vertx.vertx());
@@ -39,22 +37,21 @@ public class LoggerTest {
     assumeTrue(loader.name().equalsIgnoreCase(engineName));
   }
 
-  @Test
-  public void shouldThrowAnErrorIfFileCantBeFound() {
-    try {
-      throw new RuntimeException();
-    } catch (RuntimeException e) {
-      log.error("Oops!", e);
+  @Test(timeout = 10000)
+  public void testProcessEnv() throws Exception {
+    Object res = loader.eval("process.env");
+    Map<String, String> env;
+    if (res instanceof Map) {
+      env = (Map) res;
+    } else if (res instanceof Value) {
+      env = ((Value) res).as(Map.class);
+    } else {
+      fail("Cannot cast response object!");
+      return;
     }
-  }
 
-  @Test
-  public void shouldPrettyPrintException() {
-    Object module = loader.require("./exp.js");
-    try {
-      loader.invokeMethod(module, "a");
-    } catch (RuntimeException e) {
-      log.error("Error from Script", e);
+    for (Map.Entry<String, String> kv : env.entrySet()) {
+      System.out.println(kv.getKey() + ": " + kv.getValue());
     }
   }
 }
